% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_locf_records_.R
\name{derive_locf_records_}
\alias{derive_locf_records_}
\title{Derive LOCF (Last Observation Carried Forward) Records}
\usage{
derive_locf_records_(
  dataset,
  dataset_expected_obs,
  by_vars,
  order,
  keep_vars = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by the \code{by_vars} and the \code{order}
parameter are expected.}

\item{dataset_expected_obs}{Expected observations dataset

Data frame with all the combinations of \code{PARAMCD}, \code{PARAM}, \code{AVISIT},
\code{AVISITN}, ... which are expected in the dataset is expected.}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} those observations from \code{dataset_expected_obs}
are added to the output dataset which do not have a corresponding observation
in the input dataset or for which \code{AVAL} is NA for the corresponding observation
in the input dataset. Only variables specified in \code{by_vars} will be populated
in the newly created records.}

\item{order}{List of variables for sorting a dataset

The dataset is sorted by \code{order} before carrying the last
observation forward (eg. \code{AVAL}) within each \code{by_vars}.}

\item{keep_vars}{Variables that need carrying the last observation forward

Keep variables that need carrying the last observation forward other than \code{AVAL}
(eg. \code{PARAMN}, \code{VISITNUM}). If by default \code{NULL}, only variables specified in
\code{by_vars} and \code{AVAL} will be populated in the newly created records.}
}
\value{
The input dataset with the new "LOCF" observations added for each
\code{by_vars}. Note, a variable will only be populated in the new parameter rows
if it is specified in \code{by_vars}.
}
\description{
Adds LOCF records as new observations for each 'by group' when the dataset
does not contain observations for missed visits/time points.
}
\details{
For each group (with respect to the variables specified for the
by_vars parameter) those observations from dataset_expected_obs are added to
the output dataset
\itemize{
\item which do not have a corresponding observation in the input dataset or
\item for which \code{AVAL} is NA for the corresponding observation in the input dataset.

For the new observations, \code{AVAL} is set to the non-missing \code{AVAL} of the
previous observation in the input dataset (when sorted by \code{order}) and
\code{DTYPE} is set to "LOCF".
}
}
\author{
G Gayatri
}
\concept{der_prm_bds_findings}
\keyword{der_prm_bds_findings}
