% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_models.R
\name{efficacy_models}
\alias{efficacy_models}
\title{ANCOVA Model data processing necessary for Table 14-3.01}
\usage{
efficacy_models(data, var = NULL, wk = NULL)
}
\arguments{
\item{data}{Source dataset (filtered by flags)}

\item{var}{Variable on which model should be run}

\item{wk}{Visit to be modeled}
}
\value{
Formatted dataframe
}
\description{
This function handles the necessary data processing to handle the CDISC pilot
primary endpoint analysis. The original source can be found
\link[here]{https://github.com/atorus-research/CDISC_pilot_replication/blob/3c8e9e3798c02be8d93bd8e8944d1e0d3f6519e2/programs/funcs.R#L401}
}
