% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{esubdemo-package}
\alias{esubdemo}
\alias{esubdemo-package}
\title{esubdemo: A demo project for analysis and reporting of clinical trials}
\description{
A demo project for analysis and reporting of clinical trials.
}
\author{
\strong{Maintainer}: Yilong Zhang \email{elong0527@gmail.com}

}
\keyword{internal}
