% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_ci}
\alias{fmt_ci}
\title{Format confidence interval}
\usage{
fmt_ci(.est, .lower, .upper, digits = 2, width = digits + 3)
}
\arguments{
\item{.est}{an estimator.}

\item{.lower}{lower confidence interval bound of an estimator.}

\item{.upper}{upper confidence interval bound of an estimator.}

\item{digits}{number of digits for \code{.est}, \code{.lower}, and \code{.upper}.}

\item{width}{the total field width.}
}
\description{
Format confidence interval
}
\examples{
fmt_ci(1, -0.25, 1.32)
}
