% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_est}
\alias{fmt_est}
\title{Format point estimator}
\usage{
fmt_est(.mean, .sd, digits = c(1, 2))
}
\arguments{
\item{.mean}{mean of an estimator.}

\item{.sd}{sd of an estimator.}

\item{digits}{number of digits for \code{.mean} and \code{.sd}.}
}
\description{
Format point estimator
}
\examples{
fmt_est(1.25, 0.5)
}
