% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_pval}
\alias{fmt_pval}
\title{Format p-Value}
\usage{
fmt_pval(.p, digits = 3)
}
\arguments{
\item{.p}{a p-value.}

\item{digits}{number of digits for \code{.est}, \code{.lower}, and \code{.upper}.}
}
\description{
Format p-Value
}
\examples{
fmt_pval(0.2)
}
