% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_agegr1}
\alias{format_agegr1}
\title{Derive pooled age group}
\usage{
format_agegr1(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_agegr1} derives the \code{AGEGR1} variable, which pools subjects into age groups.
}
