% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_bmiblgr1}
\alias{format_bmiblgr1}
\title{Derive pooled baseline BMI group}
\usage{
format_bmiblgr1(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_bmiblgr1} derives the \code{BMIBLGR1} variable, which pools subjects into baseline BMI groups.
}
