% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_dcsreas}
\alias{format_dcsreas}
\title{Derive subject's disposition reason(s)}
\usage{
format_dcsreas(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_dcsreas} derives the \code{DCSREAS} variable, which contains subjects' disposition reason(s).
}
