% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_durdsgr1}
\alias{format_durdsgr1}
\title{Derive pooled disease duration group}
\usage{
format_durdsgr1(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_durdsgr1} derives the \code{DURDSGR1} variable, which pools subjects into disease duration groups.
}
