% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_eosstt}
\alias{format_eosstt}
\title{Derive subject's end of study status}
\usage{
format_eosstt(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_eosstt} derives the \code{EOSSTT} variable, which contains subject's end of study status
}
