% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_racen}
\alias{format_racen}
\title{Derive race (N)}
\usage{
format_racen(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_racen} derives the \code{RACEN} variable, which is the numeric code for \code{RACE}.
}
