% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_sitegr1}
\alias{format_sitegr1}
\title{Derive pooled site group}
\usage{
format_sitegr1(x)
}
\arguments{
\item{x}{The input vector of values to format}
}
\description{
\code{format_sitegr1} derives the \code{SITEGR1} variable, which pools sites with fewer than 3 subjects in any one treatment group.
}
