% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tplyr_helpers.R
\name{nest_rowlabels}
\alias{nest_rowlabels}
\title{Nest Row Labels in a Tplyr table}
\usage{
nest_rowlabels(.dat)
}
\arguments{
\item{.dat}{Input data set - should come from a built Tplyr table.}
}
\value{
data.frame with row labels nested
}
\description{
This is a (high ungeneralized) helper function. Current function assumes that
row_label1 groups row_label2, and turns row_label1 into a stub over its
related groups of row_label2.
}
