% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{num_fmt}
\alias{num_fmt}
\title{Number formatter}
\usage{
num_fmt(var, digits = 0, size = 10, int_len = 3)
}
\arguments{
\item{var}{Variable to format}

\item{digits}{Desired number of decimal places}

\item{size}{String size}

\item{int_len}{Space allotted for integer side of the decimal}
}
\value{
Formatted string
}
\description{
Format numbers for presentation, with proper rounding of data
}
