% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_row}
\alias{pad_row}
\title{Add a padding row below data}
\usage{
pad_row(.data, n = 1)
}
\arguments{
\item{.data}{Data to pad}

\item{n}{Number of rows to pad}
}
\value{
Dataframe with extra blank rows
}
\description{
Add a padding row below data
}
