% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pilot1wrappers-package}
\alias{pilot1wrappers}
\alias{pilot1wrappers-package}
\title{pilot1wrappers: R Consortium R Submission Pilot 1}
\description{
R Consortium R Submission Pilot 1.
}
\author{
\strong{Maintainer}: Yilong Zhang \email{elong0527@gmail.com}

Authors:
\itemize{
  \item Ning Leng
  \item Mike Stackhouse
  \item Peikun Wu
  \item Heng Wang
  \item Eli Miller
}

}
\keyword{internal}
