% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pilot3-package}
\alias{pilot3}
\alias{pilot3-package}
\title{pilot3: R Consortium R Submission Pilot 3}
\description{
R Consortium R Submission Pilot 3.
}
\author{
\strong{Maintainer}: Yilong Zhang \email{elong0527@gmail.com}

Authors:
\itemize{
  \item Ning Leng
  \item Mike Stackhouse
  \item Peikun Wu
  \item Heng Wang
  \item Eli Miller
}

Other contributors:
\itemize{
  \item Steven Haesendonckx [contributor]
}

}
\keyword{internal}
