% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pilot3-package}
\alias{pilot3}
\alias{pilot3-package}
\title{pilot3: R Consortium R Submission Pilot 3}
\description{
R Consortium R Submission Pilot 3.
}
\author{
\strong{Maintainer}: Joel Laxmana-Roche \email{laxamana.joel@gene.com}

Authors:
\itemize{
  \item Thomas Neitmann-Roche
  \item Phanikumar Tata-Syneos
  \item Robert Devine-J&J
  \item Steven Haesendonckx-J&J
  \item Rikimaru Nishimura-J&J
  \item Yutong Liu-Moderna
  \item Lei Zhao-Roche
  \item Nicole Jones-Merck
  \item Benjamin Wang-Merck
  \item Ryan Johnson-Posit
  \item Kangjie Zhang-Bayer
  \item Dadong Zhang-Illumina
  \item Benjamin Straub-GSK
  \item Declan Hodges-GSK
}

}
\keyword{internal}
