% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{round_sas}
\alias{round_sas}
\title{Round the value based on SAS rounding convention}
\usage{
round_sas(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\description{
\code{round_sas} rounds the values in its first argument to the specified number
of decimal places (default 0) (align SAS rounding convention)
}
